+++ packages/shared/src/requests.ts
@@ -14,7 +14,7 @@ export namespace GetDocumentContentRequest {
 
 export namespace GetDocumentVersionRequest {
 	export type ParamsType = vscode.TextDocumentIdentifier;
-	export type ResponseType = number | undefined;
+	export type ResponseType = number | null | undefined;
 	export type ErrorType = never;
 	export const type = new rpc.RequestType<ParamsType, ResponseType, ErrorType>('vue/docUpdated');
 }